fix: :bug: Rename hooks
